+++ tabbycat/checkins/templates/PeopleStatusMixin.vue
@@ -21,26 +21,30 @@ export default {
   },
   methods: {
     getToolTipForPerson: function (entity) {
-      let tt = `${entity.name}, a ${entity.type}`
+      if (!this.teamCodes && entity.type !== 'Team' && entity.institution === null && entity.identifier !== null) {
+        return this.gettext('%1, a %2 of no institutional affiliation with identifier of %3', entity.name, entity.type, entity.identifier[0])
+      }
+      if (!this.teamCodes && entity.type !== 'Team' && entity.institution === null) {
+        return this.gettext('%1, a %2 of no institutional affiliation with no assigned identifier', entity.name, entity.type)
+      }
+      if (!this.teamCodes && entity.type !== 'Team' && entity.identifier !== null) {
+        return this.gettext('%1, a %2 from %3 with identifier of %4', entity.name, entity.type, entity.institution.name, entity.identifier[0])
+      }
       if (!this.teamCodes && entity.type !== 'Team') {
-        if (entity.institution === null) {
-          tt += ' of no institutional affiliation'
-        } else {
-          tt += ` from ${entity.institution.name}`
-        }
+        return this.gettext('%1, a %2 from %3 with no assigned identifier', entity.name, entity.type, entity.institution.name)
       }
       if (entity.speakers !== null && entity.type === 'Team') {
-        tt += ' with speakers '
+        let speakers = []
         _.forEach(entity.speakers, (speaker) => {
-          const status = speaker.status ? 'Present; id=' : 'Absent; id='
-          tt += `${speaker.name} (${status} ${speaker.identifier[0]}) `
+          if (speaker.status) {
+            speakers.push(this.gettext('%1 (Present; id=%2)', speaker.name, speaker.identifier[0]))
+          } else {
+            speakers.push(this.gettext('%1 (Absent; id=%2)', speaker.name, speaker.identifier[0]))
+          }
         })
-      } else if (entity.identifier !== null) {
-        tt += ` with identifier of ${entity.identifier[0]}`
-      } else {
-        tt += ' with no assigned identifier '
+        return this.gettext('%1, a team with speakers %2', entity.name, speakers.join(', '))
       }
-      return tt
+      return this.gettext('%1, a %2', entity.name, entity.type)
     },
     annotatePeople: function (peopleType) {
       const self = this
@@ -58,7 +62,7 @@ export default {
       const speakers = this.annotatePeople('speakers')
       if (this.teamCodes) {
         _.forEach(speakers, (speaker) => {
-          speaker.institution = { code: 'Anonymous (due to team codes)', name: 'Anon' }
+          speaker.institution = { code: this.gettext('Anonymous (due to team codes)'), name: this.gettext('Anon') }
         })
       }
       return speakers
@@ -100,7 +104,7 @@ export default {
     annotatedAdjudicators: function () {
       _.forEach(this.adjudicators, (adjudicator) => {
         if (adjudicator.independent) {
-          adjudicator.institution = { code: 'Independent', name: 'Independent' }
+          adjudicator.institution = { code: this.gettext('Independent'), name: this.gettext('Independent') }
         }
       })
       return this.annotatePeople('adjudicators')
@@ -122,13 +126,13 @@ export default {
     peopleByInstitution: function () {
       const sortedByInstitution = _.sortBy(this.entitiesSortedByName, (p) => {
         if (p.institution === null) {
-          return 'Unaffiliated'
+          return this.gettext('Unaffiliated')
         }
         return p.institution.code
       })
       return _.groupBy(sortedByInstitution, (p) => {
         if (p.institution === null) {
-          return 'Unaffiliated'
+          return this.gettext('Unaffiliated')
         }
         return p.institution.code
       })