+++ tabbycat/printing/templates/PrintableBallotHeader.vue
@@ -2,28 +2,36 @@
   <header class="db-margins-m db-flex-row h5 mb-2 mt-3">
 
     <div class="db-flex-item">
-
-      {{ roundInfo.round }} {{ roundInfo.kind }} from {{ ballot.author }}
-      <span v-if="roundInfo.votingBallots">
-        ({{ authorPositionWithSoloCheck(ballot.authorPosition) }})
+      <span v-if="!roundInfo.votingBallots && !ballot.target"
+            v-text="gettext('%1 %2 from %3', roundInfo.round, roundInfo.kind, ballot.author)">
       </span>
-      <span v-if="ballot.target">on {{ ballot.target }}
-        <span v-if="ballot.targetPosition === 'c' || ballot.targetPosition === 'o'">(Chair)</span>
-        <span v-if="ballot.targetPosition === 'p'">(Panellist)</span>
-        <span v-if="ballot.targetPosition === 't'">(Trainee)</span>
+      <span v-if="roundInfo.votingBallots"
+            v-text="gettext('%1 %2 from %3 %4', roundInfo.round, roundInfo.kind, ballot.author,
+              authorPositionWithSoloCheck(ballot.authorPosition))">
+      </span>
+      <span v-if="ballot.target">
+        <span v-if="ballot.targetPosition === 'c' || ballot.targetPosition === 'o'"
+              v-text="gettext('%1 %2 from %3 on %4 (Chair)', roundInfo.round, roundInfo.kind, ballot.author, ballot.target)"></span>
+        <span v-if="ballot.targetPosition === 'p'"
+              v-text="gettext('%1 %2 from %3 on %4 (Panellist)', roundInfo.round, roundInfo.kind, ballot.author, ballot.target)"></span>
+        <span v-if="ballot.targetPosition === 't'"
+              v-text="gettext('%1 %2 from %3 on %4 (Trainee)', roundInfo.round, roundInfo.kind, ballot.author, ballot.target)"></span>
       </span>
 
     </div>
 
     <div v-if="ballot.venue === '' || ballot.venue === null" class="ml-auto" >
-      <span v-if="ballot.barcode">ID {{ ballot.barcode }}, </span>Venue:
+      <span v-if="ballot.barcode"
+            v-text="gettext('ID %1,', ballot.barcode)"></span>
+      <span v-text="gettext('Venue:')"></span>
       <span class="db-padding-horizontal db-fill-in"
             style="width: 232px; margin: 0 3px 0 5px; display: inline-block">
       </span>
     </div>
 
     <div v-else class="ml-auto ">
-      <span v-if="ballot.barcode">ID {{ ballot.barcode }}, </span>
+      <span v-if="ballot.barcode"
+            v-text="gettext('ID %1,', ballot.barcode)"></span>
       {{ ballot.venue.display_name }}
     </div>
 
@@ -41,17 +49,17 @@ export default {
         const panellists = _.filter(this.ballot.debateAdjudicators, da => da.position === 'p')
         if (!_.isUndefined(panellists) && panellists.length > 0) {
           const voters = _.filter(this.ballot.debateAdjudicators, da => da.position !== 't')
-          return `Chair for Panel of ${voters.length}`
+          return this.gettext('Chair for Panel of %1', voters.length)
         }
-        return 'Solo Chair'
+        return this.gettext('Solo Chair')
       } else if (position === 'o') {
-        return 'Solo Chair'
+        return this.gettext('Solo Chair')
       } else if (position === 'p') {
-        return 'Panellist'
+        return this.gettext('Panellist')
       } else if (position === 't') {
-        return 'Trainee'
+        return this.gettext('Trainee')
       } else if (position === 'TEAM') {
-        return 'Team'
+        return this.gettext('Team')
       }
       return '?'
     },