+++ tabbycat/printing/templates/PrintableFeedback.vue
@@ -4,17 +4,16 @@
     <section v-if="ballot.authorPosition === 'Team'"
              class="db-margins-m db-bordered db-flex-row db-flex-item-1">
       <div class="db-padding-horizontal db-flex-item-1 db-flex-row">
-        <div class="db-align-vertical-center db-flex-static db-vertical-center-text">
-          Did {{ ballot.target }} deliver the adjudication?
-        </div>
+        <div class="db-align-vertical-center db-flex-static db-vertical-center-text"
+             v-text="gettext('Did %1 deliver the adjudication?', ballot.target)"></div>
         <div class="db-flex-item db-flex-row">
           <div class="db-align-horizontal-center db-padding-horizontal db-align-vertical-center
                       db-flex-static db-center-text db-vertical-center-text">
-            <span class="db-padding-horizontal db-fill-in">Yes</span>
+            <span class="db-padding-horizontal db-fill-in" v-text="gettext('Yes')"></span>
           </div>
           <div class="db-align-horizontal-center db-padding-horizontal db-align-vertical-center
                       db-flex-static db-center-text db-vertical-center-text">
-            <span class="db-padding-horizontal ">No, I am submitting feedback on:</span>
+            <span class="db-padding-horizontal " v-text="gettext('No, I am submitting feedback on:')"></span>
           </div>
           <div class="db-align-horizontal-center db-fill-in db-align-vertical-center
                       db-flex-item db-center-text db-vertical-center-text">